Remove password from session mnemonic login
Mnemonic decryption is done when instantiating the sessions signer, so
does not need to be reimplemented for each session type.

This method is only used now for electrum logins, so hide it from the
top level session interface. It can be removed once ga_rust is moved to
use authenticate() for logging in.

With software login following the HW code path, auto-relogin for
software sessions is now disabled. Future changes to connection
handling (in particular, keeping state around while reconnecting)
will restore automatic reconnection for both HW and SW wallets.